Move utils folder into dataserver These aren't used elsewhere and this structure will allow proper logging
