+++ opt/scopeserver/dataserver/utils/Constant.py
